Moved ColumnLinearNoAsync module for consistency
